AudioDatabaseIndexVersion1.08.2020.html
Version 63 of the audio database index. This version improves the HTML conversion, and also changes the file name again to one that I feel could be the final name. This name change is still in use today. I copied the right encoding version this time, I might need to go back and fix the others.