Changed encoding from ANSI to UTF-8
Same as the last fix, I have changed the encoding from ANSI back to UTF-8. Garbled characters are now recovered and back to their original forms. Notepad++ does this by random, and I don't how to make it stop changing it back to ANSI every time I reopen the file (including closing and reopening the application, as the file will still be ANSI)