Don't take this out of context
This is the cover art of a album from a Vaporwave artist "Windows96" it is the only album that wasn't uploaded to YouTube, and part of the archival policy, I downloaded the album (paid for it) the cover art may seem lewd, it just came with the song. It is normal for covert art for albums to be weird.