B SECTION info.txt
The old B section divider directory, it helps me, since I can switch to a letter section by pressing almost any key on my keyboard, and it takes me to the beginning of the section, with all the song directories that start with the letter. However, this file is out of date.