ChildhoodHorr1.jpg, desktop.ini [2x]
The desktop.ini file for this directory, plus the album cover that scared 10 year old me. Another thing that contributed to this is when I was on YouTube back then, this was the only video on the site that displayed as 0:00 duration that I could find, when the duration was actually over 3 minutes. I don't know what caused that specific bug, this was a time when the iOS icon for YouTube was a TV, and when I used an iPod Touch 4th generation. Plus it was early in the morning, so there's that too.